+++ game-frame-server/game-network-message/src/main/java/com/xinyue/network/message/common/IGameError.java
@@ -2,6 +2,6 @@ package com.xinyue.network.message.common;
 
 public interface IGameError {
 
-	int getErrorCode();
+	short getErrorCode();
 	String getErrorMsg();
 }
